+++ libs/components/src/dialog/index.ts
@@ -1,0 +1,4 @@
+export * from "./dialog.component";
+export * from "./dialog-simple.component";
+export * from "./dialog.module";
+export * from "./dialog.service";
