--- libs/components/src/modal/index.ts
@@ -1,4 +1,0 @@
-export * from "./modal.component";
-export * from "./modal-simple.component";
-export * from "./modal.module";
-export * from "./modal.service";
